pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "bahai-songs-cli"
version = "0.1.1"
description = "CLI tools for the Bahá'í songs ChordPro repository"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "python-dotenv>=1.0.1",
  "requests>=2.23.0",
  "typer>=0.16.0",
]

[project.scripts]
bsp = "cli.main:app"

[tool.setuptools]
packages = ["cli"]
